% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isprime.R
\name{isprime}
\alias{isprime}
\title{judge a positive number is prime number or not}
\usage{
isprime(n)
}
\arguments{
\item{n}{a positive number}
}
\value{
T or F
}
\description{
judge a positive number is prime number or not
}
\examples{
isprime(2)
isprime(999983)


}
