% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leeLogit.R
\name{leeLogit}
\alias{leeLogit}
\title{calculate Logistic regression model's coefficients
calculate logit regression model's coefficients using Newton's or gradient optimization algorithm}
\usage{
leeLogit(y, X, method = "Newton", iter = 1000)
}
\arguments{
\item{y}{lables of samples in training data}

\item{X}{feature of samples in training data}

\item{iter}{the times we iterate in the iteration algorithm}
}
\value{
coefficients of logistic regression model
}
\description{
calculate Logistic regression model's coefficients
calculate logit regression model's coefficients using Newton's or gradient optimization algorithm
}
\examples{
leeLogit(y=c(rep(0,50),rep(1,50)),X=matrix(rnorm(100*10),ncol=10),method="Newton",iter=1000)


}
