% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newsort.R
\name{newsort}
\alias{newsort}
\title{sort a sequence of numbers from the smallest to largest.}
\usage{
newsort(a)
}
\arguments{
\item{a}{vector containing a sequence of numbers.}
}
\value{
sorted vector in which the number is posited in order.
}
\description{
sort a sequence of numbers according their size.
}
\examples{
newsort(a=c(1,7,4,3,10,-4))


}
