% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leepackage.R
\name{plotArr}
\alias{plotArr}
\title{Creates boxplots of Time of Arrival for each class}
\usage{
plotArr(x)
}
\arguments{
\item{x}{a dataset: BST 411, BST 464, or BST 494}
}
\value{
Creates a boxplot of Time of Arrival for BST 411, BST 464, or BST 494
}
\description{
Creates boxplots of Time of Arrival for each class
}

