% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leepackage.R
\name{plotDep}
\alias{plotDep}
\title{Creates boxplots of Time of Departure for each class}
\usage{
plotDep(x)
}
\arguments{
\item{x}{a dataset: BST 411, BST 464, or BST 494}
}
\value{
Creates a boxplot of Time of Departure for BST 411, BST 464, or BST 494
}
\description{
Creates boxplots of Time of Departure for each class
}

