% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power3plot.R
\name{power3plot}
\alias{power3plot}
\title{plot the power under different population mean and 3 different sample scale}
\usage{
power3plot(nseq, museq)
}
\arguments{
\item{nseq}{a vector containing three positive integer, such as c(10,20,30)}

\item{museq}{a vector containing a sequnce of number, used as the mean of nomal distribution}
}
\value{
a picture
}
\description{
plot the power change for different nomal population mean under 3 different sample scale
}
\examples{
power3plot(nseq=c(10,50,100),museq=seq(-2,2,by=0.1))


}
