% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leepackage.R
\name{sq}
\alias{sq}
\title{Square a number
Takes a numeric number and square it}
\usage{
sq(x)
}
\arguments{
\item{x}{a numeric to be squared}
}
\value{
The square of the input
}
\description{
Square a number
Takes a numeric number and square it
}

