% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leepackage.R
\name{TimeArr}
\alias{TimeArr}
\title{Takes the data and provides a summary of Time of Arrival for each class}
\usage{
TimeArr(x)
}
\arguments{
\item{x}{a dataset: BST 411, BST 464, or BST 494}
}
\value{
Mean, Median, and Standard Deviation of the Time of Arrival for each class. 1st row = mean, 2nd row = median, 3rd row = sd
}
\description{
Takes the data and provides a summary of Time of Arrival for each class
}

